#4.20 Update
Fixed shove times to 5.
Add a plugin that allows player to auto-refill.
Add training mode that makes 1vht mode more easier.
Change the jockey-skeet dmg to 170(former is 195) and fixed the hunter-skeet dmg to 150.
Change the shotgun limit to 3.
Open all talk.
Allow bot gaming.
Change the smoker self-clear dmg to 170(former is 200).